src/magnet/supervisor_agent/translator/requirements.txt
pydantic
langgraph
langgraph-cli[inmem]
langchain-core
langchain-openai
fastmcp
a2a-sdk
uvicorn
langchain-mcp-adapters
langchain-huggingface
langchain-ollama
langsmith